Inlining TikaReader inside TikaSource
