+++ sdks/java/io/tika/src/test/java/org/apache/beam/sdk/io/tika/TikaSourceTest.java
@@ -23,6 +23,7 @@ import static org.junit.Assert.assertTrue;
 import java.util.List;
 
 import org.apache.beam.sdk.coders.StringUtf8Coder;
+import org.apache.beam.sdk.io.tika.TikaSource.TikaReader;
 import org.junit.Test;
 
 /**
